refactor for caching and file clean up
