+++ frontend/src/pages/org/archived-item-qa/index.ts
@@ -1,0 +1,1 @@
+import "./archived-item-qa";
