+++ frontend/src/pages/org/archived-item-qa/ui/spinner.ts
@@ -1,0 +1,9 @@
+import { html } from "lit";
+
+export function renderSpinner() {
+  return html`<div
+    class="flex h-full w-full items-center justify-center text-2xl"
+  >
+    <sl-spinner></sl-spinner>
+  </div>`;
+}
